+++ extensions/omni_schema/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - TBD
+
+### Added
+
+* Schema dump and restore
+  functionality [#908](https://github.com/omnigres/omnigres/pull/908),[#910](https://github.com/omnigres/omnigres/pull/910)
+
 ## [0.3.0] - 2025-03-01
 
 ### Added
@@ -61,3 +68,5 @@ Initial release following a few months of iterative development.
 [0.2.3]: [https://github.com/omnigres/omnigres/pull/738]
 
 [0.3.0]: [https://github.com/omnigres/omnigres/pull/766]
+
+[0.4.0]: [https://github.com/omnigres/omnigres/pull/766]